re-factoring http client helper and adding java-docs
