--- interaction-tester/src/main/java/com/temenos/interaction/test/http/HttpUtil.java
@@ -1,22 +1,0 @@
-package com.temenos.interaction.test.http;
-
-public class HttpUtil {
-
-	public static String removeParameter(String contentType) {
-		int parameterSeparatorIndex = contentType.indexOf(";");
-		if (parameterSeparatorIndex > 0) {
-			return contentType.substring(0, parameterSeparatorIndex).trim();
-		} else {
-			return contentType;
-		}
-	}
-
-	public static String extractParameter(String contentType) {
-		int parameterSeparatorIndex = contentType.indexOf(";");
-		if (parameterSeparatorIndex > 0) {
-			return contentType.substring(parameterSeparatorIndex).trim();
-		} else {
-			return "";
-		}
-	}
-}
