+++ interaction-tester/src/main/java/com/temenos/interaction/test/internal/ActionableLink.java
@@ -1,6 +1,7 @@
 package com.temenos.interaction.test.internal;
 
 import com.temenos.interaction.test.Link;
+import com.temenos.interaction.test.Url;
 
 public interface ActionableLink extends Link {
