--- interaction-tester/src/main/java/com/temenos/interaction/test/internal/Url.java
@@ -1,21 +1,0 @@
-package com.temenos.interaction.test.internal;
-
-
-public interface Url {
-	
-	String url();
-	
-	Url baseuri(String baseUri);
-	
-	Url path(String path);
-	
-	Url queryParam(String queryParam);
-	
-	Url noPayload();
-	
-	void get();
-
-	void post();
-
-	void put();
-}
